refactor(전체): 패키지 분류 및 Main, View 클래스 생성
