+++ src/main/java/step2/Main.java
@@ -1,0 +1,10 @@
+package step2;
+
+import step2.controller.LottoMachine;
+
+public class Main {
+
+    public static void main(String[] args) {
+        LottoMachine.createLottos();
+    }
+}
