+++ src/main/java/step2/controller/LottoMachine.java
@@ -1,4 +1,8 @@
-package step2;
+package step2.controller;
+
+import step2.domain.Lotto;
+import step2.domain.LottoNumber;
+import step2.domain.Lottos;
 
 import java.util.ArrayList;
 import java.util.Collections;
