+++ src/main/java/step2/view/InputView.java
@@ -1,0 +1,4 @@
+package step2.view;
+
+public class InputView {
+}
