+++ src/test/java/step2/LottoMachineTest.java
@@ -2,9 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
+import step2.controller.LottoMachine;
 
 import static org.assertj.core.api.Assertions.assertThat;
