+++ src/test/java/step2/LottoNumberTest.java
@@ -2,6 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoNumber;
 
 import static org.assertj.core.api.Assertions.*;
