+++ src/test/java/step2/MoneyTest.java
@@ -2,6 +2,7 @@ package step2;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step2.domain.Money;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
